Keep Edit Spot live I/O under the coverage floor.
Colocate the thin Supabase store with tested editor support and drop separate repository seams so only injectable orchestration stays above the changed-line gate.
